Data Engineering/LLM/requirements.txt
flask
flask_cors
groq
psycopg2-binary==2.9.9
python-dotenv